requirements.txt
alembic==1.7.7
APScheduler==3.6.3
asgiref==3.5.2
attrs==22.1.0
cachetools==4.2.2
certifi==2021.10.8
click==8.1.3
colorama==0.4.4
coverage==6.5.0
distlib==0.3.4
Django==4.1.1
django-debug-toolbar==3.6.0
django-filter==22.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
filelock==3.7.0
greenlet==1.1.2
inflection==0.5.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.0
MarkupSafe==2.1.1
platformdirs==2.5.2
psycopg2==2.9.3
pyrsistent==0.18.1
python-status==1.0.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.42
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.0.1
uritemplate==4.1.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.1.2
